requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.5.2
beautifulsoup4==4.13.4
bidict==0.23.1
bitarray==3.3.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
cytoolz==1.0.1
distro==1.9.0
dnspython==2.7.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
eventlet==0.39.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
hexbytes==1.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mnemonic==0.21
msgspec==0.19.0
multidict==6.4.3
openai==1.78.0
packaging==25.0
parsimonious==0.10.0
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
pyunormalize==16.0.0
pywin32==310
redis==6.1.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
schedule==1.2.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
web3==7.12.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
